3. Capstone project - Dota2 draft prediction/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
numpy==1.26.3
pandas==2.2.3
xgboost==2.1.4
scikit-learn==1.6.1
joblib==1.4.2
sentence-transformers==3.4.1
torch==2.5.1
gunicorn==23.0.0
